+++ rrpproxypy/__init__.py
@@ -1,1 +1,2 @@
+from rrpproxypy.args import add_arguments
 from rrpproxypy.client import RRPproxy